Allow undo/redo of text annotation
