+++ src/components/nodes/association/association.vue
@@ -6,74 +6,18 @@
 <script>
 import joint from 'jointjs';
 import crownConfig from '@/mixins/crownConfig';
+import linkConfig from '@/mixins/linkConfig';
 import get from 'lodash/get';
-import debounce from 'lodash/debounce';
-import { validNodeColor, invalidNodeColor, defaultNodeColor } from '@/components/nodeColors';
 
 export default {
   props: ['graph', 'node', 'id', 'moddle', 'nodeRegistry'],
-  mixins: [crownConfig],
+  mixins: [crownConfig, linkConfig],
   data() {
     return {
       shape: null,
-      definition: null,
-      sourceShape: null,
-      target: null,
-      anchorPadding: 25,
     };
   },
   computed: {
-    sourceNode() {
-      return get(this.sourceShape, 'component.node');
-    },
-    targetNode() {
-      return get(this.target, 'component.node');
-    },
-    sourceConfig() {
-      return this.sourceNode && this.nodeRegistry[this.sourceNode.type];
-    },
-    targetConfig() {
-      return this.targetNode && this.nodeRegistry[this.targetNode.type];
-    },
-    elementPadding() {
-      return this.shape && this.shape.source().id === this.shape.target().id ? 20 : 1;
-    },
-  },
-  methods: {
-    beforeLoadInspector(node) {
-      return node.definition && node.definition.targetRef && node.definition.targetRef.id;
-    },
-    setBodyColor(color, target = this.target) {
-      target.attr('body/fill', color);
-      target.attr('.body/fill', color);
-    },
-    updateDefinitionLinks() {
-      const targetShape = this.shape.getTargetElement();
-      this.node.definition.targetRef = targetShape.component.node.definition;
-    },
-    completeLink() {
-      this.shape.stopListening(this.paper, 'cell:mouseleave');
-      this.shape.stopListening(this.paper, 'blank:pointerclick link:pointerclick', this.removeLink);
-      this.$emit('set-cursor', null);
-
-      this.resetPaper();
-
-      this.updateWaypoints();
-
-      const targetShape = this.shape.getTargetElement();
-      this.setBodyColor(defaultNodeColor, targetShape);
-
-      this.shape.listenTo(this.sourceShape, 'change:position', this.updateWaypoints);
-      this.shape.listenTo(targetShape, 'change:position', this.updateWaypoints);
-      this.$emit('set-cursor', 'grab');
-    },
-    updateWaypoints() {
-      const connections = this.shape.findView(this.paper).getConnection();
-      const points = connections.segments.map(segment => segment.end);
-
-      this.node.diagram.waypoint = points.map(point => this.moddle.create('dc:Point', point));
-      this.updateCrownPosition();
-    },
     isValidConnection() {
       const targetType = get(this.target, 'component.node.type');
 
@@ -101,83 +45,14 @@ export default {
 
       return true;
     },
+  },
+  methods: {
     updateRouter() {
       this.shape.router('normal', { elementPadding: this.elementPadding });
     },
-    updateLinkTarget({ clientX, clientY }) {
-      const localMousePosition = this.paper.clientToLocalPoint({ x: clientX, y: clientY });
-
-      /* Sort shapes by z-index descending; grab the shape on top (with the highest z-index) */
-      this.target = this.graph.findModelsFromPoint(localMousePosition).sort((shape1, shape2) => {
-        return shape2.get('z') - shape1.get('z');
-      })[0];
-
-      if (!this.isValidConnection()) {
-        this.$emit('set-cursor', 'not-allowed');
-        this.shape.listenToOnce(this.paper, 'blank:pointerdown link:pointerdown element:pointerdown', this.removeLink);
-
-        this.shape.target({
-          x: localMousePosition.x,
-          y: localMousePosition.y,
-        });
-
-        if (this.target) {
-          this.setBodyColor(invalidNodeColor);
-        }
-
-        return;
-      }
-
-      this.shape.stopListening(this.paper, 'blank:pointerdown link:pointerdown element:pointerdown', this.removeLink);
-
-      this.shape.target(this.target, {
-        anchor: {
-          name: this.target instanceof joint.shapes.standard.Rectangle ? 'perpendicular' : 'modelCenter',
-          args: { padding: this.anchorPadding },
-        },
-        connectionPoint: { name: 'boundary' },
-      });
-
-      this.updateRouter();
-
-      this.$emit('set-cursor', 'default');
-      this.setBodyColor(validNodeColor);
-
-      this.paper.el.removeEventListener('mousemove', this.updateLinkTarget);
-      this.shape.listenToOnce(this.paper, 'cell:pointerclick', () => {
-        this.completeLink();
-        this.updateDefinitionLinks();
-      });
-
-      this.shape.listenToOnce(this.paper, 'cell:mouseleave', () => {
-        this.paper.el.addEventListener('mousemove', this.updateLinkTarget);
-        this.shape.stopListening(this.paper, 'cell:pointerclick');
-        this.setBodyColor(defaultNodeColor);
-        this.$emit('set-cursor', 'not-allowed');
-      });
-    },
-    removeLink() {
-      this.removeShape();
-      this.resetPaper();
-    },
-    resetPaper() {
-      this.$emit('set-cursor', null);
-      this.paper.el.removeEventListener('mousemove', this.updateLinkTarget);
-      this.paper.setInteractivity(this.graph.get('interactiveFunc'));
-
-      if (this.target) {
-        this.setBodyColor(defaultNodeColor);
-      }
-    },
-  },
-  created() {
-    this.updateWaypoints = debounce(this.updateWaypoints, 100);
-  },
-  watch: {
-    target(target, previousTarget) {
-      if (previousTarget && previousTarget !== target) {
-        this.setBodyColor(defaultNodeColor, previousTarget);
-      }
+    updateDefinitionLinks() {
+      const targetShape = this.shape.getTargetElement();
+      this.node.definition.targetRef = targetShape.component.node.definition;
     },
   },
   mounted() {
@@ -197,52 +72,6 @@ export default {
         },
       },
     });
-
-    this.sourceShape = this.graph.getElements().find(element => {
-      return element.component && element.component.node.definition === this.node.definition.get('sourceRef');
-    });
-
-    this.shape.source(this.sourceShape, {
-      anchor: { name: 'modelCenter' },
-      connectionPoint: { name: 'boundary' },
-    });
-
-    this.sourceShape.embed(this.shape);
-
-    this.shape.addTo(this.graph);
-    this.shape.component = this;
-
-    const targetRef = this.node.definition.get('targetRef');
-
-    if (targetRef.id) {
-      const targetShape = this.graph.getElements().find(element => {
-        return element.component && element.component.node.definition === targetRef;
-      });
-
-      this.shape.target(targetShape, {
-        anchor: {
-          name: targetShape instanceof joint.shapes.standard.Rectangle ? 'perpendicular' : 'modelCenter',
-          args: { padding: this.anchorPadding },
-        },
-        connectionPoint: { name: 'boundary' },
-      });
-
-      this.completeLink();
-    } else {
-      this.shape.target(targetRef, {
-        connectionPoint: { name: 'boundary' },
-      });
-
-      this.paper.setInteractivity(false);
-      this.paper.el.addEventListener('mousemove', this.updateLinkTarget);
-
-      this.$emit('set-cursor', 'not-allowed');
-    }
-
-    this.updateRouter();
-  },
-  destroyed() {
-    this.updateWaypoints.cancel();
   },
 };
 </script>
